+++ src/util/http/json.go
@@ -32,21 +32,3 @@ func SendJSONOr500(log *logging.Logger, w http.ResponseWriter, m interface{}) {
 		Error500(w)
 	}
 }
-
-// SendOr404 sends an interface as JSON if its not nil (404) or fails (500)
-func SendOr404(w http.ResponseWriter, m interface{}) {
-	if m == nil {
-		Error404(w)
-	} else if SendJSON(w, m) != nil {
-		Error500(w)
-	}
-}
-
-// SendOr500 sends an interface as JSON if its not nil (500) or fails (500)
-func SendOr500(w http.ResponseWriter, m interface{}) {
-	if m == nil {
-		Error500(w)
-	} else if SendJSON(w, m) != nil {
-		Error500(w)
-	}
-}